requirements.txt
beautifulsoup4
python-dateutil
dateparser
sqlitedict
aiogram
httpx
